pyproject.toml
[project]
name = "study-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.2.3",
    "flask>=3.1.3",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "python-dotenv>=1.2.2",
]
